sui_adapter_latest::programmable_transactions
Module trace_utils

This module implements support for tracing related to PTB execution. IMPORTANT: Bodies of all public functions in this module should be enclosed in a large if statement checking if tracing is enabled or not to make sure that any errors coming from these functions only manifest itself when tracing is enabled.

Functions

trace_move_call_end

Inserts Move call end event into the trace. As is the case for all other public functions in this module, its body is (and must be) enclosed in an if statement checking if tracing is enabled.

trace_move_call_start

Inserts Move call start event into the trace. As is the case for all other public functions in this module, its body is (and must be) enclosed in an if statement checking if tracing is enabled.

trace_ptb_summary

Inserts PTB summary event into the trace. As is the case for all other public functions in this module, its body is (and must be) enclosed in an if statement checking if tracing is enabled.

trace_split_coins

Inserts split coins event into the trace. As is the case for all other public functions in this module, its body is (and must be) enclosed in an if statement checking if tracing is enabled.

trace_transfer

Inserts transfer event into the trace. As is the case for all other public functions in this module, its body is (and must be) enclosed in an if statement checking if tracing is enabled.