sui_adapter_latest::programmable_transactions::trace_utils
Function trace_split_coins

pub fn trace_split_coins(
    context: &mut ExecutionContext<'_, '_, '_>,
    trace_builder_opt: &mut Option<MoveTraceBuilder>,
    coin_type: &Type,
    input_coin: &Coin,
    split_coin_values: &[Value],
) -> Result<(), ExecutionError>

Inserts split coins event into the trace. As is the case for all other public functions in this module, its body is (and must be) enclosed in an if statement checking if tracing is enabled.