sui_adapter_latest::static_programmable_transactions::typing::invariant_checks
Module memory_safety

Functions

verify

Verifies memory safety of a transaction. This is a re-implementation of verify::memory_safety using an alternative approach given the newness of the Regex based borrow graph in that implementation. This is a set based approach were each reference is represent as a set of paths. A path is has a root (basically a T::Location plus some edge case massaging) and a list of extensions resulting from Move function calls. Each one of those Move calls gets a Delta extension for each return value. The Delta is like the “.” in the regex based implementation but where it carries a sense of identity. This identity allows for invariants from the return values of the Move call to be leveraged. For example, mutable references returned from a call cannot overlap. If we just used “.”, we would not be able to express this invariant without some sense of identity for the reference itself (which is what is going on in the Regex based implementation). This implementation stems from research work for the Move borrow checker, but would normally not be expressive enough in the presence of control flow. Luckily, PTBs do not have control flow so we can use this approach as a safety net for the Regex based implementation until that code is sufficiently. tested and hardened. Checks the following