sui_adapter_latest::static_programmable_transactions::typing::verify::private_entry_arguments

Function verify

pub fn verify<Mode: ExecutionMode>(
    env: &Env<'_, '_, '_, '_>,
    txn: &Transaction,
) -> Result<(), ExecutionError>

Checks entry taint rules. An entry function cannot have any arguments in a “hot” clique. We define cliques and “hot” with the following rules:

Each input value starts as a node in the graph, as such in its own clique.

When a value is used in a command, we “entangle” all of the values and the outputs. Conceptually, we add an edge between all of the inputs/outputs and any other nodes in the inputs’s cliques. As such, we don’t really need to track the graph as a whole, just what clique each value is in.

A value is “heats” its clique if it is a “hot potato” value, i.e. a value whose type does not have drop and does not have store.

A clique is the “hot” in one of two ways:

If it contains 1 or more “hot potato” values

If it is marked as “always hot” if a command consumes a shared object by-value

A non-public entry function cannot have any inputs that are in a hot clique.

Note that command inputs are released before checking the rules, so an entry function can consume a hot potato value if it is the last “heating” value in its clique.